Use `Set` as backing store for `Utils.Bounded`
`Set` uses balanced binary trees, so insertions are logarithmic in the size of
the set, instead of linear as in the current list-based implementation. The
`size` field keeps track of the backing set so that `Set.cardinal`, which has
run-time complexity linear in the size of the set, can be avoided

In addition, this is a much simpler implementation.

Includes a test of Bounded based on general invariants