80% may be too high of a threshold for amendments
I fear that 80% may be too high of a threshold for amending the governance system. I don't have hard evidence for this, but my sense is that even a 2/3s voting threshold is very hard to achieve (it may not be hard to achieve in small groups or groups which are very collegial, but my sense is that it is hard to achieve in large groups which have polarized factions, which i believe is very common in large groups).

What supermajority thresholds do groups tend to use (for amendments but also for other supermajority thresholds)? The most common choice seems to be 2/3 (a little over 66%), which is found in, for example, Robert's Rules, the US Constitution (although 2/3s is used for some things, US constitutional amendments must be ratified by 3/4 of the states), the German Basic Law, and elections in the Catholic Church. The next most common choice seems to be 3/5, which is found in, for example, the US Senate, the UN Security Countil, the South Korean National Assembly, the EU Commission.

A simple majority of shareholders can amend US Delaware corporations; in the UK and Germany, the threshold is 3/4 (in some cases these may be defaults that can themselves be modified by amendments) (Section 242(b)(i) of Delaware General Corporation Law http://delcode.delaware.gov/title8/c001/sc08/index.shtml ; UK Companies Act 2006 Sections 21 and 283 http://www.legislation.gov.uk/ukpga/2006/46/section/283 ; German Stock Corporation Act (Aktiengesetz) Section 179 http://www.nortonrosefulbright.com/files/german-stock-corporation-act-147035.pdf).

Thresholds higher than 3/4 are so hard to meet that they are sometimes regarded as forms of approximate consensus.

For example, the Occupy Cincinnati group used a 90% threshold for ordinary decisions, and this was said to render making decisions "pretty much impossible" (Justin Jeffre, quoted in The Occupy Movement and the Poetics of the Oppressed by Ursula McTaggart), and some experienced activists believed that high supermajority thresholds led to the Occupy Cincinnati movement's downfall (ibid).

I would recommend a threshold of 2/3s.