Improve make build-deps after a tezos-opam-repository update
First this patch makes build-deps non interactive as it should be.
More importantly, it changes the optimization problem sent to the
opam solver (and that may be a big deal).

By default the criteria is `-removed,-changed,-notuptodate`.

-changed is problematic because it means that if a dependency
has been updated in tezos-opam-repository but all tezos packages
are fine with the one currently installed, it is not changed (but
as we observe now there might still be in fact packages with
undeclared incompatible with this old version)!

-remove is also problematic because what we would like to say is
"minimize the number of removed 'root' packages" we don't care to
remove packages automatically installed and not necessary anymore
(especially when they don't compile anymore with updated version of
their deps...)

So... I reversed order in the criteria. I'm sure this is what I want
when the opam switch contains only tezos-opam-repository but
unexpected upgrades occur for devs that have done a
`opam repo add default --rank=-1`...