Clic: stop being colorless.
Writing keyword in `white` was a bad idea for people using a
black-on-white terminal...

This patch replace `white` by the default color for keywords (it
preserves the underline). This does not change the color for
white-on-block terminal (where `white` is the default color), but fix
the issue for all terminals with a clear background.